fix(hr): the exported detrend fit needs 64 bits, not 32
a_q16 is the window's DC level shifted left 16, and this sensor's raw counts
are 24-bit, so the intercept sits around 1e11. The exported copy was int32 and
wrapped: across the first six hours of real dumps fit_a spanned the entire
int32 range and 49% of the values came back negative, for a signal that is
positive by construction.

The measurement itself was never wrong -- detrend_into_work computes in int64
and only the exported copy was narrowed -- but the dump could not be inverted
back to raw counts, which was the entire point of shipping it.

The samples were fine: over 330 windows, raw >> 3 matched the paired int8 dump
on every single value, which is what localised this to the two fit fields.

Wire header 15 -> 23 bytes, 279 B per chunk, still one frame.

+++ example/get-started/dualcore/src/modules/algorithm/health/hr_autocorr.h
@@ -183,7 +183,7 @@ uint32_t hr_autocorr_total(void);
  * @return entries written, 0 if no estimate has run yet or offset is past the
  *         end of the window. */
 uint16_t hr_autocorr_last_work(int16_t *out, uint16_t offset, uint16_t n,
-                               int32_t *a_q16, int32_t *b_q16, uint8_t *shift);
+                               int64_t *a_q16, int64_t *b_q16, uint8_t *shift);
 
 /**
  * Copy the detrended window the LAST hr_autocorr_estimate() ran on, rescaled to
